api-server/requirements.txt
anyio==3.6.2
entrypoints==0.3
fastapi==0.89.1
fastapi-mqtt==1.0.7 #org 1.7
fasteners==0.14.1
gmqtt==0.6.11
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.14.0
httptools==0.1.2
httpx==0.23.3
Jinja2==3.1.2
netifaces==0.10.4
numpy==1.24.1
paho-mqtt==1.6.1
pandas==1.5.3
SQLAlchemy==2.0.0
starlette==0.22.0
uvicorn==0.20.0
uvloop==0.17.0
PyMySQL==1.0.2
requests==2.22.0
requests-unixsocket==0.2.0
simplejson==3.16.0
jsonschema==3.2.0